requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==7.0.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
asyncer==0.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
Django==3.2.8
djangorestframework==3.12.4
fastapi==0.72.0
filelock==3.6.0
filetype==1.0.9
flatbuffers==2.0
frozenlist==1.3.0
gdown==4.4.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.13.0
idna==3.3
imageio==2.18.0
llvmlite==0.36.0
multidict==6.0.2
networkx==2.8
numba==0.53.1
numpy==1.22.3
onnxruntime==1.10.0
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.0.1
promise==2.3
protobuf==3.20.1
pydantic==1.9.0
PyMatting==1.1.5
pyparsing==3.0.8
PySocks==1.7.1
python-multipart==0.0.5
pytz==2022.1
PyWavelets==1.3.0
rembg==2.0.21
requests==2.27.1
Rx==1.6.1
scikit-image==0.19.1
scipy==1.8.0
singledispatch==3.7.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
starlette==0.17.1
text-unidecode==1.3
tifffile==2022.4.28
tqdm==4.62.3
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.0
watchdog==2.1.7
whitenoise==6.0.0
yarl==1.7.2
